+++ Classes/StateWalk.cpp
@@ -12,7 +12,7 @@
 #include "GameRole.h"
 
 void StateWalk::execute(GameRole* role, EventCustom* event){
-    auto enMsgType = (int)GameRoleState::toEnum(event->getEventName());
+    auto enMsgType = GameRoleState::toEnum(event->getEventName());
     
     switch (enMsgType) {
         case GameRoleState::State::Idle:
